+++ src/adc/watermarker/__init__.py
@@ -1,1 +1,0 @@
-from ._passthrough import PassThrough